refactor: move ExceptionHandler to handers.py
